🧪 tests: consolidate and improve coverage for src/godot/headless.ts
- Merged 'headless-full.test.ts' and 'headless-security.test.ts' into 'headless.test.ts'.
- Added test cases for edge cases (null/undefined stdout, stderr, and error messages).
- Achieved 100% statement and branch coverage for 'src/godot/headless.ts'.
- Cleaned up redundant test files.
